+++ src/metabokg/app.py
@@ -12,7 +12,7 @@ Run with:
     poetry run metabokg-viz
 
 Author: Eric G. Suchanek, PhD
-Last Revision: 2026-02-28 21:25:00
+Last Revision: 2026-07-25 00:21:15
 
 """